Refactored general_utils & Fixed resizeBottomInset
