+++ lib/ui/pages/settings/widgets/dev_info.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:kanpractice/core/utils/general_utils.dart';
+import 'package:kanpractice/ui/general_utils.dart';
 import 'package:kanpractice/ui/theme/consts.dart';
 import 'package:kanpractice/ui/widgets/kp_drag_container.dart';
 import 'package:url_launcher/url_launcher.dart';
