+++ lib/ui/pages/study_modes/writing.dart
@@ -8,7 +8,7 @@ import 'package:kanpractice/ui/pages/study_modes/widgets/writing_buttons_animati
 import 'package:kanpractice/ui/widgets/kp_study_mode_app_bar.dart';
 import 'package:kanpractice/ui/widgets/canvas/kp_custom_canvas.dart';
 import 'package:kanpractice/ui/theme/consts.dart';
-import 'package:kanpractice/core/utils/general_utils.dart';
+import 'package:kanpractice/ui/general_utils.dart';
 import 'package:kanpractice/core/utils/study_modes/mode_arguments.dart';
 import 'package:kanpractice/core/utils/study_modes/study_mode_update_handler.dart';
 import 'package:kanpractice/ui/widgets/kp_learning_header_animation.dart';
