+++ lib/ui/widgets/graphs/kp_win_rate_chart.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:kanpractice/core/database/database_consts.dart';
-import 'package:kanpractice/core/utils/general_utils.dart';
+import 'package:kanpractice/ui/general_utils.dart';
 import 'package:kanpractice/ui/theme/consts.dart';
 import 'package:syncfusion_flutter_gauges/gauges.dart';
